third_party/WebKit/LayoutTests/svg/custom/immutable-properties-expected.txt

Tests whether immutable properties can not be modified.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS viewport.x is 0
PASS viewport.x is 100
PASS svgDoc.documentElement.viewport.x is 0
PASS zoomRectScreen.x is 0
PASS zoomRectScreen.x = 100; threw exception NoModificationAllowedError: Failed to set the 'x' property on 'SVGRect': The attribute is read-only..
PASS zoomEvent.previousScale is 0
PASS zoomEvent.previousScale is 0
PASS previousTranslate.x is 0
PASS previousTranslate.x = 300; threw exception NoModificationAllowedError: Failed to set the 'x' property on 'SVGPoint': The attribute is read-only..
PASS zoomEvent.previousTranslate.x is 0
PASS zoomEvent.newScale is 0
FAIL zoomEvent.newScale = 200; should throw an exception. Was 200.
PASS zoomEvent.newScale is 0
PASS newTranslate.x is 0
PASS newTranslate.x = 300; threw exception NoModificationAllowedError: Failed to set the 'x' property on 'SVGPoint': The attribute is read-only..
PASS newTranslate.x is 0
PASS successfullyParsed is true

TEST COMPLETE
